+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/reset-password/reset-password.component.ts
@@ -1,7 +1,7 @@
 import { Component, HostBinding, Input, OnInit } from '@angular/core';
 import { AuthenticatorService } from '../../../../services/authenticator.service';
 import {
-  formField,
+  FormField,
   getActorState,
   getAliasInfoFromContext,
   getFormDataFromEvent,
@@ -20,7 +20,7 @@ export class ResetPasswordComponent implements OnInit {
   public sendCodeText = translate('Send Code');
   public backToSignInText = translate('Back to Sign In');
   public labelText = translate<string>('Username');
-  public formOverrides: formField;
+  public formOverrides: FormField;
 
   constructor(public authenticator: AuthenticatorService) {}