+++ packages/angular/projects/ui-angular/src/lib/components/authenticator/components/sign-up/sign-up-form-fields/sign-up-form-fields.component.ts
@@ -2,14 +2,16 @@ import { Component, OnInit } from '@angular/core';
 import { AuthenticatorService } from '../../../../../services/authenticator.service';
 import {
   authInputAttributes,
-  formField,
-  formFieldTypes,
+  FormField,
+  FormFieldOptions,
   getActorState,
   LoginMechanism,
   SignUpAttribute,
   CommonFields,
   setFormOrder,
-  LoginMechanismArray,
+  ActorContextWithForms,
+  getActorContext,
+  translate,
 } from '@aws-amplify/ui';
 
 @Component({
@@ -24,8 +26,9 @@ export class SignUpFormFieldsComponent implements OnInit {
   public loginMechanism: LoginMechanism;
   public order: (string | number)[];
 
-  public userOverrides: formFieldTypes;
-  public formOverrides: formField;
+  public userOverrides: FormFieldOptions;
+  public formOverrides: FormField;
+  public translate = translate;
 
   constructor(private authenticator: AuthenticatorService) {}
 
@@ -58,6 +61,14 @@ export class SignUpFormFieldsComponent implements OnInit {
     ] as CommonFields[];
     this.fieldNamesCombined = [...common, ...this.fieldNames];
     this.setFormFields();
+  }
+
+  get errors() {
+    const formContext: ActorContextWithForms = getActorContext(
+      this.authenticator.authState
+    );
+    const { validationError } = formContext;
+    return validationError['password'];
   }
 
   public setFormFields() {