+++ packages/react/__tests__/exports.ts
@@ -24,6 +24,7 @@ describe('@aws-amplify/ui-react', () => {
           "CheckboxField",
           "Collection",
           "ComponentClassNames",
+          "ComponentClassObject",
           "ComponentPropsToStylePropsMap",
           "ComponentPropsToStylePropsMapKeys",
           "Divider",