+++ packages/react/src/primitives/TextField/__tests__/TextField.test.tsx
@@ -8,8 +8,7 @@ import {
   testFlexProps,
   expectFlexContainerStyleProps,
 } from '../../Flex/__tests__/Flex.test';
-import { AUTO_GENERATED_ID_PREFIX } from '../../shared/utils';
-
+import { AUTO_GENERATED_ID_PREFIX } from '../../utils/useStableId';
 const label = 'Field';
 const testId = 'testId';
 const originalWarn = console.warn;